+++ app/src/main/java/com/ceaver/assin/markets/TitleRepository.kt
@@ -5,7 +5,6 @@ import android.os.Looper
 import com.ceaver.assin.assets.AssetCategory
 import com.ceaver.assin.database.Database
 import com.ceaver.assin.threading.BackgroundThreadExecutor
-import java.math.BigDecimal
 import java.util.*
 
 object TitleRepository {
@@ -139,7 +138,7 @@ object TitleRepository {
         return unionList
     }
 
-    fun lookupPriceAsync(symbol: Title, reference: Title, callbackInMainThread: Boolean, callback: (Optional<BigDecimal>) -> Unit) {
+    fun lookupPriceAsync(symbol: Title, reference: Title, callbackInMainThread: Boolean, callback: (Optional<Double>) -> Unit) {
         BackgroundThreadExecutor.execute {
             val price = lookupPrice(symbol, reference)
             if (callbackInMainThread)
@@ -149,7 +148,7 @@ object TitleRepository {
         }
     }
 
-    fun lookupPrice(symbol: Title, reference: Title): Optional<BigDecimal> {
+    fun lookupPrice(symbol: Title, reference: Title): Optional<Double> {
         if (reference.symbol == "EUR" || reference.symbol == "CHF") {
             TODO("not yet implemented")
         }
